Add clear_process_command_for_buildpacks migration
This is a migration to clear process.command when it matches its app's
droplet's detected_start_command. When the two are equivalent, we assume
that the command has not been explicitly set by the user. If the command
has not been explicitly set by the user, we want to defer to the
buildpack's detected start command which we do by nulling out
process.command.

[#157756113]

Signed-off-by: Davis W. Frank <dwfrank@pivotal.io>